Add files via upload
Delighted to present my latest Java mini project - a dynamic JSP application with a seamless blend of JavaBeans, HTML, CSS, and JDBC! 🌐💻

This project's focal point is the authentication system, emphasizing secure logins and logouts. Leveraging JavaBeans, I've crafted a robust backend, while the front end, elegantly designed with HTML and CSS, ensures a user-friendly experience.

The JSP (Java Servlet Pages) here is used for creating a dynamic web content that links our frontend with the backend.

As this project is done in VS code, you can see the video is also provided for the reference and understanding more about the mini project. 🎥

The JDBC (Java Database Connectivity) integration is a key highlight, providing a streamlined connection between our Java application and the Oracle database. Specifically, the login system validates customer credentials by cross-referencing the input with a dedicated table in the database. Upon successful verification, the system retrieves and displays the customer's order details from the database.

This project showcases the synergy of Java technologies, HTML, CSS, and database management, making it a comprehensive exploration of web development. Excited to share the journey and insights gained through this endeavor!